TINY-11009: Moved files around and implemented link selection logic
