+++ modules/tinymce/src/plugins/link/main/ts/ui/Controls.ts
@@ -1,58 +1,95 @@
 import { Fun, Optional, Optionals } from '@ephox/katamari';
 
 import Editor from 'tinymce/core/api/Editor';
-import { InlineContent } from 'tinymce/core/api/ui/Ui';
+import { NodeChangeEvent } from 'tinymce/core/api/EventTypes';
+import { InlineContent, Menu, Toolbar } from 'tinymce/core/api/ui/Ui';
 
 import * as Options from '../api/Options';
 import * as Actions from '../core/Actions';
-import * as Selection from '../core/Selection';
+import * as OpenLink from '../core/OpenLink';
 import * as Utils from '../core/Utils';
 
-const setupButtons = (editor: Editor, linkSelection: Selection.LinkSelection): void => {
+const openDialog = (editor: Editor) => (): void => {
+  editor.execCommand('mceLink', false, { dialog: true });
+};
+
+const toggleState = (editor: Editor, toggler: (e: NodeChangeEvent) => void): () => void => {
+  editor.on('NodeChange', toggler);
+  return () => editor.off('NodeChange', toggler);
+};
+
+const toggleLinkState = (editor: Editor) => (api: Toolbar.ToolbarToggleButtonInstanceApi | Menu.ToggleMenuItemInstanceApi): () => void => {
+  const updateState = () => {
+    api.setActive(!editor.mode.isReadOnly() && Utils.isInAnchor(editor, editor.selection.getNode()));
+    api.setEnabled(editor.selection.isEditable());
+  };
+  updateState();
+  return toggleState(editor, updateState);
+};
+
+const toggleLinkMenuState = (editor: Editor) => (api: Menu.MenuItemInstanceApi): () => void => {
+  const updateState = () => {
+    api.setEnabled(editor.selection.isEditable());
+  };
+  updateState();
+  return toggleState(editor, updateState);
+};
+
+const toggleRequiresLinkState = (editor: Editor) => (api: Toolbar.ToolbarButtonInstanceApi | Menu.MenuItemInstanceApi): () => void => {
+  const hasLinks = (parents: Node[]) => Utils.hasLinks(parents) || Utils.hasLinksInSelection(editor.selection.getRng());
+  const parents = editor.dom.getParents(editor.selection.getStart());
+  const updateEnabled = (parents: Node[]) => {
+    api.setEnabled(hasLinks(parents) && editor.selection.isEditable());
+  };
+  updateEnabled(parents);
+  return toggleState(editor, (e) => updateEnabled(e.parents));
+};
+
+const setupButtons = (editor: Editor, openLink: OpenLink.LinkSelection): void => {
   editor.ui.registry.addToggleButton('link', {
     icon: 'link',
     tooltip: 'Insert/edit link',
     shortcut: 'Meta+K',
-    onAction: Actions.openDialog(editor),
-    onSetup: Actions.toggleLinkState(editor)
+    onAction: openDialog(editor),
+    onSetup: toggleLinkState(editor)
   });
 
   editor.ui.registry.addButton('openlink', {
     icon: 'new-tab',
     tooltip: 'Open link',
-    onAction: Actions.gotoSelectedLink(editor, linkSelection),
-    onSetup: Actions.toggleRequiresLinkState(editor)
+    onAction: openLink.gotoSelectedLink,
+    onSetup: toggleRequiresLinkState(editor)
   });
 
   editor.ui.registry.addButton('unlink', {
     icon: 'unlink',
     tooltip: 'Remove link',
-    onAction: () => Utils.unlink(editor),
-    onSetup: Actions.toggleRequiresLinkState(editor)
+    onAction: () => Actions.unlink(editor),
+    onSetup: toggleRequiresLinkState(editor)
   });
 };
 
-const setupMenuItems = (editor: Editor, linkSelection: Selection.LinkSelection): void => {
+const setupMenuItems = (editor: Editor, openLink: OpenLink.LinkSelection): void => {
   editor.ui.registry.addMenuItem('openlink', {
     text: 'Open link',
     icon: 'new-tab',
-    onAction: Actions.gotoSelectedLink(editor, linkSelection),
-    onSetup: Actions.toggleRequiresLinkState(editor)
+    onAction: openLink.gotoSelectedLink,
+    onSetup: toggleRequiresLinkState(editor)
   });
 
   editor.ui.registry.addMenuItem('link', {
     icon: 'link',
     text: 'Link...',
     shortcut: 'Meta+K',
-    onAction: Actions.openDialog(editor),
-    onSetup: Actions.toggleLinkMenuState(editor)
+    onAction: openDialog(editor),
+    onSetup: toggleLinkMenuState(editor)
   });
 
   editor.ui.registry.addMenuItem('unlink', {
     icon: 'unlink',
     text: 'Remove link',
-    onAction: () => Utils.unlink(editor),
-    onSetup: Actions.toggleRequiresLinkState(editor)
+    onAction: () => Actions.unlink(editor),
+    onSetup: toggleRequiresLinkState(editor)
   });
 };
 
@@ -71,7 +108,7 @@ const setupContextMenu = (editor: Editor): void => {
   });
 };
 
-const setupContextToolbars = (editor: Editor, linkSelection: Selection.LinkSelection): void => {
+const setupContextToolbars = (editor: Editor, openLink: OpenLink.LinkSelection): void => {
   const collapseSelectionToEnd = (editor: Editor) => {
     editor.selection.collapse(false);
   };
@@ -104,7 +141,7 @@ const setupContextToolbars = (editor: Editor, linkSelection: Selection.LinkSelec
       type: 'contextformtogglebutton',
       icon: 'link',
       tooltip: 'Link',
-      onSetup: Actions.toggleLinkState(editor)
+      onSetup: toggleLinkState(editor)
     },
     label: 'Link',
     predicate: (node) => Options.hasContextToolbar(editor) && Utils.isInAnchor(editor, node),
@@ -122,13 +159,13 @@ const setupContextToolbars = (editor: Editor, linkSelection: Selection.LinkSelec
           const node = editor.selection.getNode();
           // TODO: Make a test for this later.
           buttonApi.setActive(Utils.isInAnchor(editor, node));
-          return Actions.toggleLinkState(editor)(buttonApi);
+          return toggleLinkState(editor)(buttonApi);
         },
         onAction: (formApi) => {
           const value = formApi.getValue();
           const text = getLinkText(value);
           const attachState = { href: value, attach: Fun.noop };
-          Utils.link(editor, attachState, {
+          Actions.link(editor, attachState, {
             href: value,
             text,
             title: Optional.none(),
@@ -147,7 +184,7 @@ const setupContextToolbars = (editor: Editor, linkSelection: Selection.LinkSelec
         onSetup: onSetupLink,
         // TODO: The original inlite action was quite complex. Are we missing something with this?
         onAction: (formApi) => {
-          Utils.unlink(editor);
+          Actions.unlink(editor);
           formApi.hide();
         }
       },
@@ -157,7 +194,7 @@ const setupContextToolbars = (editor: Editor, linkSelection: Selection.LinkSelec
         tooltip: 'Open link',
         onSetup: onSetupLink,
         onAction: (formApi) => {
-          Actions.gotoSelectedLink(editor, linkSelection)();
+          openLink.gotoSelectedLink();
           formApi.hide();
         }
       }
@@ -166,12 +203,12 @@ const setupContextToolbars = (editor: Editor, linkSelection: Selection.LinkSelec
 };
 
 const setup = (editor: Editor): void => {
-  const linkSelection = Selection.setup(editor);
+  const openLink = OpenLink.setup(editor);
 
-  setupButtons(editor, linkSelection);
-  setupMenuItems(editor, linkSelection);
+  setupButtons(editor, openLink);
+  setupMenuItems(editor, openLink);
   setupContextMenu(editor);
-  setupContextToolbars(editor, linkSelection);
+  setupContextToolbars(editor, openLink);
 };
 
 export {
