+++ modules/tinymce/src/plugins/link/main/ts/ui/Dialog.ts
@@ -4,6 +4,7 @@ import Editor from 'tinymce/core/api/Editor';
 import { Dialog } from 'tinymce/core/api/ui/Ui';
 
 import * as Options from '../api/Options';
+import * as Actions from '../core/Actions';
 import { ListOptions } from '../core/ListOptions';
 import * as Utils from '../core/Utils';
 import { DialogChanges } from './DialogChanges';
@@ -15,7 +16,7 @@ const handleSubmit = (editor: Editor, info: LinkDialogInfo) => (api: Dialog.Dial
   const data: LinkDialogData = api.getData();
 
   if (!data.url.value) {
-    Utils.unlink(editor);
+    Actions.unlink(editor);
     // Temporary fix. TODO: TINY-2811
     api.close();
     return;
@@ -40,7 +41,7 @@ const handleSubmit = (editor: Editor, info: LinkDialogInfo) => (api: Dialog.Dial
   };
 
   DialogConfirms.preprocess(editor, changedData).then((pData) => {
-    Utils.link(editor, attachState, pData);
+    Actions.link(editor, attachState, pData);
   });
 
   api.close();
